requirements.txt
Flask==3.0.0
google-cloud-discoveryengine
google-cloud-aiplatform
gunicorn==21.2.0
Flask-Cors==4.0.0
